requirements.txt
asgiref==3.2.10
Django==3.0.10
django-debug-toolbar==3.1.1
Pillow==7.2.0
psycopg2==2.8.6
pytz==2020.1
qrcode==6.1
six==1.15.0
sqlparse==0.3.1
